Create data_header_updater.py
This script is to change the feature-names of a big data file;

The problem was:
I have a friend uses SAS to process big and datasets from health insurance companies. 
Many feature-names are very long and have spaces, so SAS cannot read it.
I wrote this script to replace the first line with new feature names and copy the rest data, 
which is an very fast and efficient way to solve his problem.